Correlation Heat Map with or Without Dendrogram

Calculates Pearson Correlation from a Matrix and plots resulting correlation values on a heatmap with color key, histogram, and optional dendrograms.

Usage

corrheatmap(
  data,
  colors = c("#00ffff", "#ff00ff", "#ffff00"),
  n_colors = 20,
  title = "Correlation Heatmap",
  dend = "both"
)

Arguments

data

a matrix or a data frame

title

(optional) a character vector. Default is "Correlation Heatmap"

dend

(optional) a string: "both","row","column","none". Default is both.

color

(optional) a vector of three colors. Default is cyan, magenta, and yellow.

Value

A heatmap graph in a new quartz window.

Examples

#Graph will open in a new Quartz window when the corrheatmap function is run
#Here are some examples of how to use the function
corrheatmap(mtcars, title="Correlations of Car Features", dend="row")
corrheatmap(mtcars, title="Correlations of Car Features", dend="col")
corrheatmap(mtcars, title="Correlations of Car Features", dend="both")
corrheatmap(mtcars, title="Correlations of Car Features", dend="none")