Interactive Network Plot

Create interactive network plot for a data frame of edge list format. Designed for exploratory data analysis.

Usage

interact_network_plot(
  data,
  initial_pt = 1,
  target_pt = 2,
  linkDistance = 10,
  linkColour = "#666",
  nodeColour = "#69b3a2",
  opacity = 0.9,
  title = "Network Graph",
  subtitle = "Exploring Interactions and Relationships"
)

Arguments

data

a data frame of edge list format. This means each row should specify a link between 2 nodes. There should be two columns; one for the inital point, and a second target point.

initial_pt

column NUMBER of initial point variable. Default is 1

target_pt

column NUMBER of target point variable. Default is 2

linkDistance

numeric distance between the links in pixels

linkColour

color of edges, MUST be a common color for the whole graph. default is #666

nodeColour

color of nodes, MUST be a common color for the whole graph. Default is #69b3a2

opacity

opacity of nodes. 0=transparent. 1=no transparency. Default is .9

title

default is "Network Graph"

subtitle

default is "Exploring Interactions and Relationships"

Value

an interactive network graph

Examples

interact_network_plot(network_data_test)

  
Network Graph

  
Exploring Interactions and Relationships

#Same as example above because default column numbers are 1 and 2
interact_network_plot(network_data_test, initial_pt=1, target_pt = 2)

  
Network Graph

  
Exploring Interactions and Relationships

interact_network_plot(network_data_test, initial_pt = 3, target_pt = 4)

  
Network Graph

  
Exploring Interactions and Relationships